Lots of design has been done
I'm thoroughly satisfied with how the v0.1 design came out. The sidebar
looks great and is very functional. The reading experience is
phenomenal. It's hard for me to got back to the original SEP after
reading on Aristotl. Now I just need to figure out Capistrano, build in
some robust error checking and other things, and then I will be
confident to put this out there for more people to see.